Fixed Colour Strings, Did some other stuff
Made it so that the colour references were all a reference to a pair and colour so that the colours had names.

Found that the rows and columns were the wrong way around so I switched those.

Added the settings button row, although the functionality has yet to be implemented.

Added the score to the confirm screen.