Refactoring and Got rid of top bar
I refactored the MainMenuScreen so that all the stuff to do with navigation went into a new file called NavigationManager and made it so that there was a Routes file which stored information on the name of the Routes. Also got rid of the top bar meaning I'll have to alter the layouts of some of the screens a bit later.